876: Renamed ordersApi to ordersHelpers, fixed ordersSlice tests
