+++ src/features/orders/ordersActions.ts
@@ -58,7 +58,7 @@ import {
   requestOrders,
   takeOrder,
   withdrawETH,
-} from "./ordersApi";
+} from "./ordersHelpers";
 import {
   setErrors,
   setOrders,
